backend/requirements.txt
Django>=4.2,<6
djangorestframework>=3.15.0,<4
supabase>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
opencv-python>=4.9.0
numpy>=1.26.0
Pillow>=10.0.0
requests>=2.31.0